# apps/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio API
  version: 1.0.0
  description: |
    The implemented HTTP API for the portfolio application. All successful responses use
    `{ success: true, data, meta }`; `meta` includes `requestId` and `timestamp`.
    Errors use `{ success: false, error: { code, message, details? } }`.
servers:
  - url: /api/v1
...
paths:
...
  /profile:
    get:
      tags: [Public]
      summary: Portfolio profile snapshot
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '503': { $ref: '#/components/responses/Error' }
  /projects:
    get:
      tags: [Public]
      summary: Published portfolio projects
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '503': { $ref: '#/components/responses/Error' }
  /projects/{slug}:
    get:
      tags: [Public]
      summary: Published project case study by slug
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
...
  /admin/profile:
    get:
      tags: [Admin content]
      summary: Read the portfolio profile
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [Admin content]
      summary: Create or replace the portfolio profile
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Admin content]
      summary: Update the portfolio profile
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: ADMIN role required for every `/admin/*` route and `/auth/me`. }
...
  schemas:
    SuccessEnvelope: { type: object, required: [success, data, meta], properties: { success: { type: boolean, enum: [true] }, data: {}, meta: { $ref: '#/components/schemas/Meta' } } }
    Meta: { type: object, properties: { requestId: { type: string }, timestamp: { type: string, format: date-time }, page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer }, hasNextPage: { type: boolean }, hasPreviousPage: { type: boolean } } }
    ErrorEnvelope: { type: object, required: [success, error], properties: { success: { type: boolean, enum: [false] }, error: { type: object, required: [code, message], properties: { code: { type: string }, message: { type: string }, details: {} } } } }
...
  responses:
    Success: { description: Successful response, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } } }
...
    Error: { description: Structured application error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...